+++ includes/class-dotstudiopro-api-activator.php
@@ -39,19 +39,22 @@ class Dotstudiopro_Api_Activator {
         // create the ECPT metabox database table
         if ($wpdb->get_var("show tables like '$dsp_video_table'") != $dsp_video_table) {
             $sql = "CREATE TABLE " . $dsp_video_table . " (
-		`id` bigint(20) NOT NULL AUTO_INCREMENT,
-		`video_id` longtext NOT NULL,
-		`video_detail` longtext NULL,
-		UNIQUE KEY id (id)
-		);";
+        		`id` bigint(20) NOT NULL AUTO_INCREMENT,
+        		`video_id` longtext NOT NULL,
+        		`video_detail` longtext NULL,
+        		UNIQUE KEY id (id)
+        		);";
 
             require_once(ABSPATH . 'wp-admin/includes/upgrade.php');
             dbDelta($sql);
         }
 
         $message = sprintf(wp_kses(__('You need to enter your API Key in order to use its features <a href="%s">Do so here.</a>', 'dotstudiopro-api'), array('a' => array('href' => array()))), esc_url(admin_url() . 'admin.php?page=dsp-api-settings'));
-
         $dsp_admin->add_admin_notice($message);
+        $dsp_admin->import_default_content();
+        if ( ! is_network_admin() && ! $dsp_admin->dsp_get_param( 'activate-multi' ) ) {
+            set_transient( 'dsp_plugin_welcome_redirect', 1, 30 );
+        }
     }
 
 }